requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
torch==2.2.1
tqdm==4.66.2
transformers==4.40.0
wandb==0.16.6
accelerate==0.29.3